Moved styles inside sidebar folder, removed unnecessary classnames
